backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
openai
stripe
pyjwt
requests
python-dotenv
